LayoutTests/imported/w3c/web-platform-tests/shadow-dom/reference-target/tentative/aria-labelledby-expected.txt

  Label 2

PASS Label 1
PASS Label 2
PASS Label 3
FAIL Label from host Label 1 assert_equals: <input class="ex" aria-labelledby="x-label4" data-expectedlabel="Label from host Label 1"> expected "Label from host Label 1" but got ""
PASS aria-labelledby
